[FL-3598] Fix PVS-Studio warnings in NFC refactoring (#3091)
* Fix PVS warnings in furi_hal_nfc_timer.c

* Fix PVS warnings in crypto1.c

* Fix PVS warnings in dict_attack.c

* Fix PVS warnings in felica_poller.c

* Fix PVS warnings in furi_hal_nfc.c

* Fix PVS warnings in furi_hal_nfc_event.c

* Fix PVS warnings in iso14443_3a_poller.c

* Fix PVS warnings in iso14443_3b_poller.c

* Fix PVS warnings in iso15693_3_poller.c

* Fix PVS warnings in nfc_protocol_support.c

* Fix PVS warnings in nfc_scanner.c

* Fix PVS warnings in opal.c

* Fix PVS warnings in mf_ultralight.c

* Fix PVS warnings in nfc_supported_cards.c

* Fix PVS warnings in nfc_scene_mf_ultralight_unlock_warn.c

* Fix PVS warnings in nfc_scene_mf_classic_mfkey_nonces_info.c

* Fix PVS warnings in mf_ultralight_poller_i.c

* Fix PVS warnings in mf_ultralight_listener_i.c

* Fix some PVS warnings in mf_ultralight_listener.c

* Fix PVS warnings in nfc_scene_mf_classic_dict_attack.c

* Fix PVS warnings in mf_classic_poller_i.c

* Fix some PVS warnings in mf_classic_listener.c

* Fix some PVS warnings in bit_buffer.c

* Fix PVS warnings in mf_classic_key_cache.c

* Fix PVS warnings in nfc_data_generator.c

* Fix some PVS warnings in mf_classic.c

* nfc: fix PVS warnings

* mf ultralight: fix tearing flag data manipulation

* nfc: fix build

---------

+++ firmware/targets/f7/furi_hal/furi_hal_nfc_event.c
@@ -36,7 +36,7 @@ FuriHalNfcEvent furi_hal_nfc_wait_event_common(uint32_t timeout_ms) {
                                                                              timeout_ms;
     uint32_t event_flag =
         furi_thread_flags_wait(FURI_HAL_NFC_EVENT_INTERNAL_ALL, FuriFlagWaitAny, event_timeout);
-    if(event_flag != FuriFlagErrorTimeout) {
+    if(event_flag != (unsigned)FuriFlagErrorTimeout) {
         if(event_flag & FuriHalNfcEventInternalTypeIrq) {
             furi_thread_flags_clear(FuriHalNfcEventInternalTypeIrq);
             FuriHalSpiBusHandle* handle = &furi_hal_spi_bus_handle_nfc;
